+++ src/core/series/types.ts
@@ -20,6 +20,8 @@ import type {
     HeatmapSeriesData,
     LineSeries,
     LineSeriesData,
+    LineSeriesDataLabelPlacementFallback,
+    LineSeriesDataLabelPlacementPosition,
     LineSeriesInterpolation,
     LineSeriesLineBaseStyle,
     LinearGradient,
@@ -292,6 +294,8 @@ export type PreparedLineSeries = {
         allowOverlap: boolean;
         html: boolean;
         format?: ValueFormat;
+        placement: 'auto' | LineSeriesDataLabelPlacementPosition[];
+        placementFallback: LineSeriesDataLabelPlacementFallback;
     };
     marker: {
         states: {